Neatened things up
