+++ backend.server/src/main/java/com/clog/Clog/LogDetailFiles/LogDetailRepository.java
@@ -1,9 +1,9 @@
-package com.clog.Clog;
+package com.clog.Clog.LogDetailFiles;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LogEventRepository extends JpaRepository<LogEvent, String>, JpaSpecificationExecutor<LogEvent>{
+public interface LogDetailRepository extends JpaRepository<LogDetail, String>, JpaSpecificationExecutor<LogDetail>{
 
 } 
-
