+++ backend.server/src/main/java/com/clog/Clog/LogEventFiles/LogEvent.java
@@ -1,6 +1,5 @@
-package com.clog.Clog;
+package com.clog.Clog.LogEventFiles;
 
-import java.sql.Date;
 import java.sql.Timestamp;
 
 import javax.persistence.Entity;
@@ -9,17 +8,14 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-
 @Entity
 @Table(name= "LOG_DETAILS_T")
-public class LogDetail {
+public class LogEvent {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
     private String global_instance_id;
     private String business_domain;
     private String business_subdomain;
-    private String version;
-    private String local_instance_id;
     private String eai_transaction_id;
     private String eai_domain;
     private String hostname;
@@ -29,11 +25,9 @@ public class LogDetail {
     private String severity;
     private String priority;
     private Timestamp creation_time;
-    private String reasoning_scope;
-    private String process_id;
     private String category_name;
     private String activity;
-    private String msg;
+
     public String getGlobal_instance_id() {
         return global_instance_id;
     }
@@ -51,18 +45,6 @@ public class LogDetail {
     }
     public void setBusiness_subdomain(String business_subdomain) {
         this.business_subdomain = business_subdomain;
-    }
-    public String getVersion() {
-        return version;
-    }
-    public void setVersion(String version) {
-        this.version = version;
-    }
-    public String getLocal_instance_id() {
-        return local_instance_id;
-    }
-    public void setLocal_instance_id(String local_instance_id) {
-        this.local_instance_id = local_instance_id;
     }
     public String getEai_transaction_id() {
         return eai_transaction_id;
@@ -118,18 +100,6 @@ public class LogDetail {
     public void setCreation_time(Timestamp creation_time) {
         this.creation_time = creation_time;
     }
-    public String getReasoning_scope() {
-        return reasoning_scope;
-    }
-    public void setReasoning_scope(String reasoning_scope) {
-        this.reasoning_scope = reasoning_scope;
-    }
-    public String getProcess_id() {
-        return process_id;
-    }
-    public void setProcess_id(String process_id) {
-        this.process_id = process_id;
-    }
     public String getCategory_name() {
         return category_name;
     }
@@ -142,11 +112,6 @@ public class LogDetail {
     public void setActivity(String activity) {
         this.activity = activity;
     }
-    public String getMsg() {
-        return msg;
-    }
-    public void setMsg(String msg) {
-        this.msg = msg;
-    }
+    
     
 }
