+++ backend.server/src/main/java/com/clog/Clog/LogEventFiles/LogEventRepository.java
@@ -1,9 +1,9 @@
-package com.clog.Clog;
-
+package com.clog.Clog.LogEventFiles;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LogDetailRepository extends JpaRepository<LogDetail, String>, JpaSpecificationExecutor<LogDetail>{
+public interface LogEventRepository extends JpaRepository<LogEvent, String>, JpaSpecificationExecutor<LogEvent>{
 
 } 
+
