+++ backend.server/src/main/java/com/clog/Clog/LogEventFiles/LogEventsSearchCriteria.java
@@ -1,4 +1,4 @@
-package com.clog.Clog;
+package com.clog.Clog.LogEventFiles;
 
 import java.sql.Timestamp;
